address review: split run.go into checkpoint/restore/spec files
run.go was ~1060 lines. Split the file (bowei) into cohesive units in the same
package, no logic changes:
- checkpoint.go: CheckpointWorkload + listFiles + teardownActor
- restore.go:    RestoreWorkload + rewriteSnapshotSocketPaths + repointActorRootfsDisk
- spec.go:       ensureKataCompatibleSpec + defaultKataMounts + defaultKataResources
run.go keeps RunWorkload + the shared boot/agent/net helpers (~500 lines).
