+++ cmd/ateom-microvm/run.go
@@ -18,14 +18,12 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log/slog"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"runtime"
-	"strings"
 	"time"
 
 	"github.com/agent-substrate/substrate/cmd/ateom-microvm/internal/ch"
@@ -33,7 +31,6 @@ import (
 	"github.com/agent-substrate/substrate/cmd/ateom-microvm/internal/third_party/kata/agentpb"
 	"github.com/agent-substrate/substrate/internal/ateompath"
 	"github.com/agent-substrate/substrate/internal/proto/ateompb"
-	"github.com/opencontainers/runtime-spec/specs-go"
 	"golang.org/x/sys/unix"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -491,567 +488,6 @@ func waitForFile(path string, d time.Duration) bool {
 		}
 		time.Sleep(50 * time.Millisecond)
 	}
-}
-
-// ensureKataCompatibleSpec augments the bundle's config.json with the fields
-// kata's OCI conversion requires but atelet's (gVisor-oriented) spec omits.
-// Without linux.resources, kata's ContainerConfig nil-derefs and the shim
-// crashes. This shaper is a bridge; a future atelet change should emit
-// runtime-appropriate specs so it can retire.
-func ensureKataCompatibleSpec(bundle, id, netnsPath string) (*specs.Spec, error) {
-	specPath := filepath.Join(bundle, "config.json")
-	b, err := os.ReadFile(specPath)
-	if err != nil {
-		return nil, fmt.Errorf("reading %q: %w", specPath, err)
-	}
-	var spec specs.Spec
-	if err := json.Unmarshal(b, &spec); err != nil {
-		return nil, fmt.Errorf("parsing %q: %w", specPath, err)
-	}
-
-	if spec.Linux == nil {
-		spec.Linux = &specs.Linux{}
-	}
-	if spec.Linux.Resources == nil {
-		spec.Linux.Resources = defaultKataResources()
-	}
-	if spec.Linux.CgroupsPath == "" {
-		spec.Linux.CgroupsPath = "/ateomchv/" + id
-	}
-
-	// atelet's spec carries gVisor pause-model CRI annotations
-	// (container-type=container, sandbox-id=pause). kata reads those and waits
-	// for a separate "pause" sandbox that we never create, failing with "the
-	// sandbox hasn't been created". Strip them so kata treats this single
-	// container as its own sandbox (creates the VM), as in the integration tests.
-	for k := range spec.Annotations {
-		if strings.HasPrefix(k, "io.kubernetes.cri.") {
-			delete(spec.Annotations, k)
-		}
-	}
-
-	// NB: no virtio-fs-overlay annotation here. With the STOCK shim, this spec is
-	// for the "carrier" container that only boots the VM + shares the RO base over
-	// virtio-fs. ateom assembles the actual overlay rootfs itself by driving the
-	// kata-agent CreateContainer over ttrpc (see RunWorkload) — no patched shim.
-
-	// Point the network namespace at our interior netns (which holds the pod's
-	// eth0); kata finds eth0 there and wires it to the VM's virtio-net.
-	netnsSet := false
-	for i := range spec.Linux.Namespaces {
-		if spec.Linux.Namespaces[i].Type == specs.NetworkNamespace {
-			spec.Linux.Namespaces[i].Path = netnsPath
-			netnsSet = true
-		}
-	}
-	if !netnsSet {
-		spec.Linux.Namespaces = append(spec.Linux.Namespaces, specs.LinuxNamespace{
-			Type: specs.NetworkNamespace, Path: netnsPath,
-		})
-	}
-
-	// Replace atelet's gVisor-oriented mounts (minimal /dev tmpfs, a
-	// /etc/resolv.conf host bind that ENOENTs against the distroless rootfs) with
-	// the exact set `ctr run --runtime io.containerd.kata.v2` emits, which kata's
-	// agent accepts. (Static shaper; pod DNS integration is future work.)
-	//
-	// KNOWN GAP vs the gVisor runtime: this also drops atelet's read-only actor
-	// identity bind mount (/run/ate/actor-id). The micro-VM guest can't see host
-	// paths (the rootfs is a virtio-blk disk, not a shared filesystem), and
-	// reset-to-golden restores guest RAM + rootfs from the golden snapshot, so a
-	// per-actor file written into the rootfs would be shadowed/incorrect on restore.
-	// Exposing the identity needs a per-actor volume injected from OUTSIDE the golden
-	// state; not yet implemented. No micro-VM workload depends on it today.
-	spec.Mounts = defaultKataMounts()
-
-	out, err := json.MarshalIndent(&spec, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshaling spec: %w", err)
-	}
-	if err := os.WriteFile(specPath, out, 0o600); err != nil {
-		return nil, fmt.Errorf("writing %q: %w", specPath, err)
-	}
-	return &spec, nil
-}
-
-// defaultKataMounts mirrors the mount set `ctr run --runtime io.containerd.kata.v2`
-// produces (the proven-good shape for the kata agent).
-func defaultKataMounts() []specs.Mount {
-	return []specs.Mount{
-		{Destination: "/proc", Type: "proc", Source: "proc", Options: []string{"nosuid", "noexec", "nodev"}},
-		{Destination: "/dev", Type: "tmpfs", Source: "tmpfs", Options: []string{"nosuid", "strictatime", "mode=755", "size=65536k"}},
-		{Destination: "/dev/pts", Type: "devpts", Source: "devpts", Options: []string{"nosuid", "noexec", "newinstance", "ptmxmode=0666", "mode=0620", "gid=5"}},
-		{Destination: "/dev/shm", Type: "tmpfs", Source: "shm", Options: []string{"nosuid", "noexec", "nodev", "mode=1777", "size=65536k"}},
-		{Destination: "/dev/mqueue", Type: "mqueue", Source: "mqueue", Options: []string{"nosuid", "noexec", "nodev"}},
-		{Destination: "/sys", Type: "sysfs", Source: "sysfs", Options: []string{"nosuid", "noexec", "nodev", "ro"}},
-		{Destination: "/run", Type: "tmpfs", Source: "tmpfs", Options: []string{"nosuid", "strictatime", "mode=755", "size=65536k"}},
-	}
-}
-
-// defaultKataResources mirrors the device allowlist + cpu shares that
-// `ctr run --runtime io.containerd.kata.v2` emits (the proven-good shape).
-func defaultKataResources() *specs.LinuxResources {
-	dev := func(t string, major, minor int64, access string) specs.LinuxDeviceCgroup {
-		d := specs.LinuxDeviceCgroup{Allow: true, Type: t, Access: access}
-		if major != 0 {
-			d.Major = &major
-		}
-		if minor >= 0 {
-			d.Minor = &minor
-		}
-		return d
-	}
-	shares := uint64(1024)
-	return &specs.LinuxResources{
-		Devices: []specs.LinuxDeviceCgroup{
-			{Allow: false, Access: "rwm"},
-			dev("c", 1, 3, "rwm"),    // /dev/null
-			dev("c", 1, 8, "rwm"),    // /dev/random
-			dev("c", 1, 7, "rwm"),    // /dev/full
-			dev("c", 5, 0, "rwm"),    // /dev/tty
-			dev("c", 1, 5, "rwm"),    // /dev/zero
-			dev("c", 1, 9, "rwm"),    // /dev/urandom
-			dev("c", 5, 1, "rwm"),    // /dev/console
-			dev("c", 136, -1, "rwm"), // pts
-			dev("c", 5, 2, "rwm"),    // /dev/ptmx
-		},
-		CPU: &specs.LinuxCPU{Shares: &shares},
-	}
-}
-
-// CheckpointWorkload suspends the actor and writes a portable CH snapshot.
-//
-// Contract with atelet (mirrors ateom-gvisor): after we return, atelet uploads
-// the checkpoint dir to object storage, then tears down bundles and resets the
-// actor dir.
-//
-// ateom drives the ateom-owned CH's REST api-socket: pause -> snapshot
-// file://<CheckpointStateDir> (config.json + state.json + sparse memory-ranges) ->
-// tear the VMM down. The actor's rootfs lives on the host-backed /dev/vdb, not a
-// guest tmpfs overlay-upper, so the snapshot is naturally memory-only and small —
-// no RAM-backed upper to wipe and no balloon to inflate before snapshot.
-func (s *AteomService) CheckpointWorkload(ctx context.Context, req *ateompb.CheckpointWorkloadRequest) (*ateompb.CheckpointWorkloadResponse, error) {
-	s.lock.Lock()
-	defer s.lock.Unlock()
-
-	ns := req.GetActorTemplateNamespace()
-	name := req.GetActorTemplateName()
-	id := req.GetActorId()
-
-	s.actorLogger.EmitLifecycleLog("Actor checkpointing", id, name, ns)
-
-	// The actor's CH was booted by RunWorkload or relaunched by RestoreWorkload;
-	// either way ateom owns it and tracks its api-socket.
-	ra := s.running[id]
-	chSocket := kata.CLHSocketPath(id)
-	if ra != nil && ra.apiSocket != "" {
-		chSocket = ra.apiSocket
-	}
-	client := ch.NewClient(chSocket)
-	if err := client.WaitReady(ctx, 10*time.Second); err != nil {
-		return nil, fmt.Errorf("while waiting for CH api-socket: %w", err)
-	}
-
-	tPause := time.Now()
-	if err := client.Pause(ctx); err != nil {
-		return nil, fmt.Errorf("while pausing guest: %w", err)
-	}
-	dPause := time.Since(tPause)
-
-	checkpointDir := ateompath.CheckpointStateDir(ns, name, id)
-	// Start from a clean dir so CH's snapshot files are the only contents.
-	if err := os.RemoveAll(checkpointDir); err != nil {
-		return nil, fmt.Errorf("while clearing checkpoint dir %q: %w", checkpointDir, err)
-	}
-	if err := os.MkdirAll(checkpointDir, 0o700); err != nil {
-		return nil, fmt.Errorf("while creating checkpoint dir %q: %w", checkpointDir, err)
-	}
-
-	// Record the FROZEN base id (the id the guest's virtio-fs find-paths are pinned
-	// to, <baseID>/rootfs). For a cold (owned-boot) actor this is its own id; for a
-	// restored actor it is the golden id propagated via ra.baseID (set from the
-	// snapshot we restored from). RestoreWorkload reads this to lay the
-	// reconstructed-from-image base at the path the guest expects. We can NOT derive
-	// it from config.json (its socket paths get rewritten to the current id on every
-	// restore, losing the invariant golden id).
-	baseID := id
-	if ra != nil && ra.baseID != "" {
-		baseID = ra.baseID
-	}
-	if err := os.WriteFile(filepath.Join(checkpointDir, baseIDFile), []byte(baseID), 0o600); err != nil {
-		return nil, fmt.Errorf("while writing %s: %w", baseIDFile, err)
-	}
-
-	// NB: the snapshot is memory-only (config/state/memory-ranges + base-id). The RO
-	// base (/dev/vda) and the writable rootfs (/dev/vdb, reset to golden at restore)
-	// are reconstructed on every node, so neither ships in the snapshot.
-
-	slog.InfoContext(ctx, "Snapshotting guest", slog.String("id", id), slog.String("dir", checkpointDir))
-	tSnapshot := time.Now()
-	if err := client.Snapshot(ctx, checkpointDir); err != nil {
-		return nil, fmt.Errorf("while snapshotting guest: %w", err)
-	}
-	dSnapshot := time.Since(tSnapshot)
-
-	// Diff-snapshot completion for an OnDemand-restored actor: CH's snapshot here is
-	// sparse — only the pages faulted in since the OnDemand restore — so on its own
-	// it's INCOMPLETE (the un-faulted pages were being demand-paged from the restore
-	// source). Overlay it onto that source to rebuild a COMPLETE memory-ranges, so the
-	// snapshot is self-contained and re-restorable. (A cold-run actor has no restore
-	// source and its snapshot is already complete — no merge.)
-	if ra != nil && ra.restoreSourceDir != "" {
-		base := filepath.Join(ra.restoreSourceDir, "memory-ranges")
-		delta := filepath.Join(checkpointDir, "memory-ranges")
-		tMerge := time.Now()
-		// Reuse base's on-disk working set (rename + overlay) instead of copying it —
-		// CH is paused and about to be torn down, and base is discarded after. See
-		// MergeDeltaIntoBase. (Falls back to the copying merge across filesystems.)
-		if err := ch.MergeDeltaIntoBase(ctx, base, delta); err != nil {
-			return nil, fmt.Errorf("while merging OnDemand delta into restore source: %w", err)
-		}
-		slog.InfoContext(ctx, "Merged OnDemand delta into base (complete snapshot)",
-			slog.String("id", id), slog.Duration("merge", time.Since(tMerge)))
-	}
-
-	// reset-to-golden support: save the actor's /dev/vdb AS-OF this (paused,
-	// consistent) snapshot as a verbatim golden template, so future restores can
-	// recreate the disk byte-identical to what the snapshot's guest RAM expects
-	// while discarding the actor's later rootfs writes. Saved once (the first/golden
-	// checkpoint) and kept; best-effort (without it, restore reopens the live disk =
-	// continuity). TODO: ship the template with the snapshot for cross-node restore
-	// (it's golden, shipped once per template, like the OCI base).
-	actorDir := ateompath.ActorPath(ns, name, id)
-	if tmpl := filepath.Join(actorDir, goldenRootfsDiskName); fileMissing(tmpl) {
-		if cerr := copyDiskFile(ctx, filepath.Join(actorDir, actorRootfsDiskName), tmpl); cerr != nil {
-			slog.WarnContext(ctx, "Failed to save golden rootfs template; restore will reopen live disk", slog.Any("err", cerr))
-		} else {
-			slog.InfoContext(ctx, "Saved golden rootfs disk template", slog.String("id", id))
-		}
-	}
-
-	// Report exactly the files we wrote so atelet ships precisely the CH snapshot
-	// (config.json + state.json + memory-ranges + base-id), not gVisor's fixed set.
-	// Memory-only: the RO base is reconstructed from the OCI image at restore.
-	snapshotFiles, err := listFiles(checkpointDir)
-	if err != nil {
-		return nil, fmt.Errorf("while listing snapshot files: %w", err)
-	}
-
-	// Tear down: the actor returns to "available". Best-effort; the snapshot is
-	// already on disk for atelet to ship.
-	tTeardown := time.Now()
-	s.teardownActor(ctx, id, ra, client)
-	dTeardown := time.Since(tTeardown)
-	delete(s.running, id)
-
-	// Tear down the per-activation actor network (mirrors gVisor).
-	if err := s.cleanupActorNetwork(ctx); err != nil {
-		slog.WarnContext(ctx, "Failed to clean up actor network after checkpoint", slog.Any("err", err))
-	}
-
-	s.actorLogger.EmitLifecycleLog("Actor checkpointed", id, name, ns)
-	slog.InfoContext(ctx, "Actor checkpointed", slog.String("id", id), slog.Any("snapshot_files", snapshotFiles),
-		slog.Duration("pause", dPause),
-		slog.Duration("snapshot", dSnapshot), slog.Duration("teardown", dTeardown))
-	return &ateompb.CheckpointWorkloadResponse{SnapshotFiles: snapshotFiles}, nil
-}
-
-// listFiles returns the (relative) names of regular files directly under dir.
-func listFiles(dir string) ([]string, error) {
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		return nil, err
-	}
-	var files []string
-	for _, e := range entries {
-		if e.Type().IsRegular() {
-			files = append(files, e.Name())
-		}
-	}
-	return files, nil
-}
-
-// teardownActor stops the ateom-owned CH VMM for an actor. Best-effort: the
-// snapshot is already on disk, so this only needs to release resources. ra may be
-// nil (e.g. ateom restarted and lost in-memory state).
-func (s *AteomService) teardownActor(ctx context.Context, id string, ra *runningActor, client *ch.Client) {
-	if client != nil {
-		tShutdown := time.Now()
-		shutCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-		if err := client.Shutdown(shutCtx); err != nil {
-			slog.WarnContext(ctx, "CH shutdown failed (continuing teardown)", slog.Any("err", err))
-		}
-		cancel()
-		slog.InfoContext(ctx, "CH API shutdown done", slog.Duration("took", time.Since(tShutdown)))
-	}
-
-	if ra != nil {
-		// Close the kata-agent client kept open for stdout/stderr forwarding. This
-		// fails the forwarding goroutines' in-flight ReadStdout/ReadStderr calls, so
-		// they return io.EOF and exit (no goroutine leak). Guarded so a second
-		// teardown / a never-forwarded actor is a no-op.
-		if ra.logAgent != nil {
-			_ = ra.logAgent.Close()
-			ra.logAgent = nil
-		}
-
-		// Kill the CH process ateom launched.
-		if ra.chCmd != nil && ra.chCmd.Process != nil {
-			_ = ra.chCmd.Process.Kill()
-			_, _ = ra.chCmd.Process.Wait()
-		}
-	}
-
-	// Sweep any leftover per-sandbox host-side state + orphaned per-sandbox
-	// processes. This is ateom's own cleanup (process kill + unmount + rm).
-	kata.CleanupSandboxState(ctx, id)
-}
-
-// RestoreWorkload restores the actor on a (possibly different) pod by relaunching
-// cloud-hypervisor directly from the downloaded snapshot and resuming.
-//
-// Contract with atelet: the memory-only snapshot dir (config.json + state.json +
-// memory-ranges + base-id) has been downloaded to RestoreStateDir.
-//
-// There is NO virtiofsd and NO shared-dir to reconstruct — the rootfs is the
-// writable /dev/vdb disk, which CH reopens from the path recorded in the snapshot
-// config.json. Steps: rewrite the vsock socket path to this actor's VMDir,
-// reset /dev/vdb to the golden disk template (or rebuild it from the OCI image),
-// rebuild the tap (the snapshot's virtio-net is fd-backed → fresh net_fds),
-// relaunch CH with --restore, and resume. Guest RAM (incl. the actor's in-memory
-// state and frozen network config) comes back from the memory-only snapshot.
-func (s *AteomService) RestoreWorkload(ctx context.Context, req *ateompb.RestoreWorkloadRequest) (resp *ateompb.RestoreWorkloadResponse, retErr error) {
-	s.lock.Lock()
-	defer s.lock.Unlock()
-
-	ns := req.GetActorTemplateNamespace()
-	name := req.GetActorTemplateName()
-	id := req.GetActorId()
-	restoreDir := ateompath.RestoreStateDir(ns, name, id)
-	tStart := time.Now()
-
-	s.actorLogger.EmitLifecycleLog("Actor restoring", id, name, ns)
-
-	rr := s.resolveRuntime(req.GetRuntimeAssetPaths())
-	kata.CleanupSandboxState(ctx, id)
-
-	// Repoint the snapshot's vsock socket to this actor's VMDir (the disk + kernel
-	// paths are content-addressed/per-actor and already line up on the same node).
-	if err := rewriteSnapshotSocketPaths(restoreDir, id); err != nil {
-		return nil, fmt.Errorf("while rewriting snapshot socket paths: %w", err)
-	}
-	srcID := id
-	if b, rerr := os.ReadFile(filepath.Join(restoreDir, baseIDFile)); rerr == nil {
-		if v := strings.TrimSpace(string(b)); v != "" {
-			srcID = v
-		}
-	}
-	if err := os.MkdirAll(kata.VMDir(id), 0o700); err != nil {
-		return nil, fmt.Errorf("while creating VM dir: %w", err)
-	}
-
-	// Recreate the /dev/vdb backing file the snapshot references (the actor dir),
-	// reset-to-golden. Two ways, both byte-consistent with the golden snapshot's
-	// guest ext4 cache:
-	//   - same-node: a verbatim golden template (copyDiskFile) — guaranteed identical.
-	//   - cross-node: rebuild from the OCI image atelet unpacked to the bundle at
-	//     restore (mkfs.ext4 -d is LAYOUT-deterministic for identical inputs, so the
-	//     data blocks land at the same offsets the guest cache expects; only the
-	//     superblock UUID/timestamps differ, which are cached in RAM and not re-read).
-	// Either way the actor's prior rootfs writes are discarded (gVisor semantics).
-	containers := req.GetSpec().GetContainers()
-	if len(containers) != 1 {
-		return nil, status.Errorf(codes.Unimplemented, "ateom-microvm supports exactly one container, got %d", len(containers))
-	}
-	actorDir := ateompath.ActorPath(ns, name, id)
-	diskPath := filepath.Join(actorDir, actorRootfsDiskName)
-	if tmpl := filepath.Join(actorDir, goldenRootfsDiskName); !fileMissing(tmpl) {
-		if err := copyDiskFile(ctx, tmpl, diskPath); err != nil {
-			return nil, fmt.Errorf("while resetting rootfs disk to golden (template): %w", err)
-		}
-		slog.InfoContext(ctx, "Reset actor rootfs disk to golden (template)", slog.String("id", id))
-	} else {
-		bundleRootfs := filepath.Join(ateompath.OCIBundlePath(ns, name, id, containers[0].GetName()), "rootfs")
-		if err := kata.BuildExt4Image(ctx, bundleRootfs, diskPath); err != nil {
-			return nil, fmt.Errorf("while reconstructing rootfs disk from image: %w", err)
-		}
-		slog.InfoContext(ctx, "Reconstructed actor rootfs disk from image", slog.String("id", id))
-	}
-
-	// Repoint the snapshot config's writable /dev/vdb disk at THIS actor's
-	// reconstructed backing file. The golden snapshot recorded the golden actor's
-	// per-actor disk path, which is stale on any pod restoring a different actor
-	// (and absent on any node that never ran the golden) — unlike /dev/vda, the
-	// content-addressed kata image whose path is identical on every node.
-	if err := repointActorRootfsDisk(restoreDir, diskPath); err != nil {
-		return nil, fmt.Errorf("while repointing actor rootfs disk in snapshot config: %w", err)
-	}
-
-	// Networking: rebuild the per-activation veth + tap; the snapshot's virtio-net
-	// is fd-backed, so CH needs fresh tap FDs (net_fds) on restore.
-	if err := s.setupActorNetwork(ctx); err != nil {
-		return nil, fmt.Errorf("while setting up actor network: %w", err)
-	}
-	defer func() {
-		if retErr != nil {
-			if cleanupErr := s.cleanupActorNetwork(ctx); cleanupErr != nil {
-				slog.WarnContext(ctx, "Failed to clean up actor network after Restore failure", slog.Any("err", cleanupErr))
-			}
-		}
-	}()
-	netDevs, err := ch.SnapshotNetDevices(restoreDir)
-	if err != nil {
-		return nil, fmt.Errorf("while reading snapshot net devices: %w", err)
-	}
-	var restoredNets []ch.RestoredNet
-	var tapFiles []*os.File
-	defer func() {
-		for _, f := range tapFiles {
-			_ = f.Close()
-		}
-	}()
-	for i, nd := range netDevs {
-		files, terr := s.setupRestoreTap(ctx, fmt.Sprintf("tap%d_kata", i), nd.QueuePairs)
-		if terr != nil {
-			return nil, fmt.Errorf("while building restore tap for %s: %w", nd.ID, terr)
-		}
-		tapFiles = append(tapFiles, files...)
-		rn := ch.RestoredNet{ID: nd.ID}
-		for _, f := range files {
-			rn.FDs = append(rn.FDs, int(f.Fd()))
-		}
-		restoredNets = append(restoredNets, rn)
-	}
-
-	// Relaunch CH and restore with the tap FDs attached (SCM_RIGHTS). CH reopens
-	// /dev/vda (image) + /dev/vdb (actor rootfs) from the snapshot config paths.
-	apiSocket := filepath.Join(kata.VMDir(id), "clh-api-restore.sock")
-	chCmd, client, err := ch.LaunchVMM(ctx, ch.LaunchVMMOptions{
-		Binary: rr.chBinary, APISocket: apiSocket, Stdout: slogWriter{ctx}, Stderr: slogWriter{ctx},
-	})
-	if err != nil {
-		return nil, fmt.Errorf("while launching VMM for restore: %w", err)
-	}
-	defer func() {
-		if retErr != nil && chCmd.Process != nil {
-			_ = chCmd.Process.Kill()
-		}
-	}()
-	// OnDemand (userfaultfd) memory restore: ~75ms vs ~1.8s eager, and it keeps the
-	// memfd SPARSE so the next suspend isn't the eager-copy-densified full-RAM scan.
-	// CH's OnDemand snapshot alone would be INCOMPLETE (it writes only faulted pages,
-	// dropping the un-faulted ones it demand-pages from this source) — so
-	// CheckpointWorkload overlays CH's delta onto this source (restoreSourceDir) to
-	// rebuild a complete snapshot. CH demand-pages from restoreDir for the VM's whole
-	// lifetime, so it must persist until teardown (atelet keeps it until reset).
-	if err := client.RestoreWithNetFDs(ctx, restoreDir, restoredNets, "OnDemand"); err != nil {
-		return nil, fmt.Errorf("while restoring VM with net FDs: %w", err)
-	}
-	if err := client.Resume(ctx); err != nil {
-		return nil, fmt.Errorf("while resuming restored guest: %w", err)
-	}
-
-	ra := &runningActor{chCmd: chCmd, apiSocket: apiSocket, baseID: srcID, restoreSourceDir: restoreDir}
-
-	// Re-attach stdout/stderr forwarding: the restored guest's container + kata-agent
-	// are alive, so a fresh dial over this actor's vsock resumes ReadStdout/ReadStderr
-	// (same containerID==execID==id as the cold run). Best-effort — a failed dial must
-	// not fail the restore (the actor is already running); forwarding is just skipped.
-	vsockPath := kata.VsockSocketPath(id)
-	logAC, dialErr := dialAgentRetry(ctx, vsockPath, 15*time.Second)
-	if dialErr != nil {
-		slog.WarnContext(ctx, "post-restore agent dial failed; actor log forwarding disabled for this restore",
-			slog.String("id", id), slog.Any("err", dialErr))
-	} else {
-		ra.logAgent = logAC
-		s.startActorLogForwarding(logAC, id, name, ns, containers[0].GetName())
-	}
-
-	s.running[id] = ra
-	s.actorLogger.EmitLifecycleLog("Actor restored", id, name, ns)
-	slog.InfoContext(ctx, "Actor restored (owned-boot, virtio-blk rootfs)",
-		slog.String("id", id), slog.Duration("total", time.Since(tStart)))
-	return &ateompb.RestoreWorkloadResponse{}, nil
-}
-
-// rewriteSnapshotSocketPaths repoints the snapshot config.json's per-sandbox
-// hybrid-vsock socket from the source actor's VMDir to the restoring actor's
-// VMDir, so the socket we create is the one CH reopens. The kernel and /dev/vda
-// kata image are content-addressed static files with identical paths on every
-// node, so they need no rewrite; the writable /dev/vdb actor rootfs disk is
-// per-actor and is repointed separately (see repointActorRootfsDisk).
-func rewriteSnapshotSocketPaths(snapshotDir, id string) error {
-	cfgPath := filepath.Join(snapshotDir, "config.json")
-	b, err := os.ReadFile(cfgPath)
-	if err != nil {
-		return err
-	}
-	var cfg map[string]any
-	if err := json.Unmarshal(b, &cfg); err != nil {
-		return fmt.Errorf("parsing %q: %w", cfgPath, err)
-	}
-	if vsock, ok := cfg["vsock"].(map[string]any); ok {
-		vsock["socket"] = kata.VsockSocketPath(id)
-	}
-	// The owned-boot path captures the guest serial console to a file under the
-	// source actor's VMDir (Serial{Mode:"File"}). On restore that path is stale
-	// (points at the golden/source pod's VMDir), so CH's CreateConsoleDevice fails
-	// (No such file or directory). Repoint it at this actor's VMDir.
-	if serial, ok := cfg["serial"].(map[string]any); ok {
-		if mode, _ := serial["mode"].(string); mode == "File" {
-			serial["file"] = filepath.Join(kata.VMDir(id), "serial.log")
-		}
-	}
-	out, err := json.Marshal(cfg)
-	if err != nil {
-		return err
-	}
-	if err := os.WriteFile(cfgPath, out, 0o600); err != nil {
-		return err
-	}
-	return nil
-}
-
-// repointActorRootfsDisk rewrites the snapshot config.json so the writable
-// /dev/vdb actor rootfs disk points at this actor's reconstructed backing file
-// (diskPath). The actor rootfs disk lives under the actor's per-actor directory
-// (keyed by actor id), so the golden snapshot's recorded path is the GOLDEN
-// actor's — stale on any pod restoring a different actor, and absent on any node
-// that never ran the golden. (This is the disk analogue of the serial.file
-// repoint in rewriteSnapshotSocketPaths.) The disk is identified by basename so
-// the read-only /dev/vda kata image (a content-addressed static file) is left
-// untouched; it is an error if no actor rootfs disk is present to repoint.
-func repointActorRootfsDisk(snapshotDir, diskPath string) error {
-	cfgPath := filepath.Join(snapshotDir, "config.json")
-	b, err := os.ReadFile(cfgPath)
-	if err != nil {
-		return err
-	}
-	var cfg map[string]any
-	if err := json.Unmarshal(b, &cfg); err != nil {
-		return fmt.Errorf("parsing %q: %w", cfgPath, err)
-	}
-	rewrote := false
-	if disks, ok := cfg["disks"].([]any); ok {
-		for _, d := range disks {
-			dm, ok := d.(map[string]any)
-			if !ok {
-				continue
-			}
-			if p, _ := dm["path"].(string); filepath.Base(p) == actorRootfsDiskName {
-				dm["path"] = diskPath
-				rewrote = true
-			}
-		}
-	}
-	if !rewrote {
-		return fmt.Errorf("no %q disk found in %q to repoint", actorRootfsDiskName, cfgPath)
-	}
-	out, err := json.Marshal(cfg)
-	if err != nil {
-		return err
-	}
-	return os.WriteFile(cfgPath, out, 0o600)
 }
 
 // slogWriter adapts an io.Writer to slog at info level, capturing the
